requirements.txt
-e git+https://github.com/kevinastone/django-api-rest-and-angular.git@b0a1ea487c6d7028e491e7558a7a0cfd2243d2ec#egg=API_Example
astroid==1.6.1
backports.functools-lru-cache==1.5
bootstrap-admin==0.3.7.1
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.4
django-ckeditor==5.3.0
django-extensions==1.8.1
django-js-asset==0.1.1
django-map-widgets==0.1.9
django-oauth-toolkit==1.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-sslserver==0.20
djangorestframework==3.6.3
drf-nested-routers==0.90.0
enum34==1.1.6
et-xmlfile==1.0.1
futures==3.2.0
idna==2.5
isort==4.3.4
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
lazy-object-proxy==1.3.1
lml==0.0.1
lxml==4.1.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.4.9
Pillow==4.2.1
psycopg2==2.7.3
pyexcel-io==0.5.4
pyexcel-xlsx==0.5.4
pylint==1.8.2
python-docx==0.8.6
pytz==2017.2
requests==2.18.3
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.11.0
sorl-thumbnail==12.3
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
